requirements.txt
langchain>=0.3,<0.4
langchain-community>=0.3,<0.4
langchain-openai>=0.3,<0.4
langchain-text-splitters>=0.3,<0.4
langchain-pinecone>=0.2,<0.3
openai>=1.0,<2.0
pypdf>=4,<6
pinecone>=5,<7
tiktoken>=0.7,<1.0
python-dotenv>=1.0,<2.0


